LayoutTests/fast/dom/DOMException/prototype-object-expected.txt

This tests the prototype chain of DOMException objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.toString() is "Error: HierarchyRequestError: DOM Exception 3"
PASS Object.prototype.toString.call(e) is "[object DOMException]"
PASS Object.prototype.toString.call(e.__proto__) is "[object DOMExceptionPrototype]"
PASS Object.prototype.toString.call(e.__proto__.__proto__) is "[object Object]"
PASS e.constructor.toString() is "function DOMException() {\n    [native code]\n}"
PASS e instanceof DOMException is true
PASS e instanceof Error is true
PASS e.constructor is window.DOMException
PASS e.HIERARCHY_REQUEST_ERR is e.constructor.HIERARCHY_REQUEST_ERR
PASS e.HIERARCHY_REQUEST_ERR is 3
PASS e.code is 3
PASS e.name is "HierarchyRequestError"
PASS e.message is "HierarchyRequestError: DOM Exception 3"
PASS successfullyParsed is true

TEST COMPLETE
